Allow mplexporter to be unbundled.
