+++ vispy/ext/_bundled/_mpl_py3k_compat.py
